Move Error layout file name
